requirements.txt
pyyaml
openai-whisper
ffmpeg-python
torch
tqdm>=4.65.0